Optimized Further
So... not sure what I did. I was testing it with the line:

pinString = (int)Serial.read();

and it wasn't working... but the numbers were always 48 higher then I
wanted them... NO clue why. After subtracting 48, tested it with every
pin number we need (0-7) and it's working, so I commented out the old
code just in case the improved method doesn't work when testing.