Bugfix
It's a bit of a sloppy fix, the first command that is input on Python
will result in "?Ready!" being shown, but the second one will work
correctly. It no longer queues back the commands, so:

1:1;
?Ready!

0:1;
Pin 1 turning off.

as opposed to previously it would turn pin one on because of the first
command being that.