Extremely buggy web functionality added
Basically unusable, just a lot of foundation work. Bugfix coming soon,
along with a bit of code cleanup. All the components we need I THINK are
there, but they're calling each other at the wrong times, and there is a
lot of unneeded code that's clogging things up a bit.