Added run.sh
To start the program now, type ./run.sh in a system shell. If it doesn't
work because of pemissions, change permissions so you can execute it. If
it's a different error, then do it the old way with:

python -m SimpleHTTPServer

and

python CientModuleProxyv3.py

If there are problems binding the server to a port, make sure you aren't
running any other programs listening on port 8000.